+++ crates/storage/src/structured_storage.rs
@@ -44,7 +44,6 @@ use crate::{
     StorageSize,
     StorageWrite,
 };
-use core::ops::Deref;
 
 #[cfg(feature = "std")]
 use std::{
@@ -382,7 +381,7 @@ where
         self.inner
             .get(key_bytes.as_ref(), <M as TableWithBlueprint>::column())
             // TODO: Return `Value` instead of cloned `Vec<u8>`.
-            .map(|value| value.map(|value| value.deref().clone()))
+            .map(|value| value.map(|value| value.to_vec()))
     }
 }