+++ crates/storage/src/transactional.rs
@@ -468,7 +468,7 @@ where
         self.changes
             .entry(column.id())
             .or_default()
-            .insert(k, WriteOperation::Insert(Value::new(buf.to_vec())));
+            .insert(k, WriteOperation::Insert(Value::from(buf)));
         Ok(buf.len())
     }
 
@@ -661,7 +661,7 @@ mod test {
             let storage = InMemoryStorage::<Column>::default();
             let mut view = storage.read_transaction();
             let key = vec![0xA, 0xB, 0xC];
-            let expected = Arc::new(vec![1, 2, 3]);
+            let expected = Value::from([1, 2, 3]);
             view.put(&key, Column::Metadata, expected.clone()).unwrap();
             // test
             let ret = view.get(&key, Column::Metadata).unwrap();
@@ -675,7 +675,7 @@ mod test {
             let mut storage = InMemoryStorage::<Column>::default();
             let mut write = storage.write_transaction();
             let key = vec![0xA, 0xB, 0xC];
-            let expected = Arc::new(vec![1, 2, 3]);
+            let expected = Value::from([1, 2, 3]);
             write.put(&key, Column::Metadata, expected.clone()).unwrap();
             write.commit().unwrap();
 
@@ -692,7 +692,7 @@ mod test {
             let mut storage = InMemoryStorage::<Column>::default();
             let mut write = storage.write_transaction();
             let key = vec![0xA, 0xB, 0xC];
-            let expected = Arc::new(vec![1, 2, 3]);
+            let expected = Value::from([1, 2, 3]);
             write.put(&key, Column::Metadata, expected.clone()).unwrap();
             write.commit().unwrap();
 
@@ -713,13 +713,13 @@ mod test {
             // setup
             let mut storage = InMemoryStorage::<Column>::default();
             let mut write = storage.write_transaction();
-            let expected = Arc::new(vec![1, 2, 3]);
+            let expected = Value::from([1, 2, 3]);
             write
                 .put(&[0xA, 0xB, 0xC], Column::Metadata, expected.clone())
                 .unwrap();
             // test
             let ret = write
-                .replace(&[0xA, 0xB, 0xC], Column::Metadata, Arc::new(vec![2, 4, 6]))
+                .replace(&[0xA, 0xB, 0xC], Column::Metadata, Value::from([2, 4, 6]))
                 .unwrap();
             // verify
             assert_eq!(ret, Some(expected))
@@ -731,7 +731,7 @@ mod test {
             let mut storage = InMemoryStorage::<Column>::default();
             let mut write = storage.write_transaction();
             let key = vec![0xA, 0xB, 0xC];
-            let expected = Arc::new(vec![1, 2, 3]);
+            let expected = Value::from([1, 2, 3]);
             write.put(&key, Column::Metadata, expected.clone()).unwrap();
             // test
             let ret = write.take(&key, Column::Metadata).unwrap();
@@ -747,7 +747,7 @@ mod test {
             let mut storage = InMemoryStorage::<Column>::default();
             let mut write = storage.write_transaction();
             let key = vec![0xA, 0xB, 0xC];
-            let expected = Arc::new(vec![1, 2, 3]);
+            let expected = Value::from([1, 2, 3]);
             write.put(&key, Column::Metadata, expected.clone()).unwrap();
             write.commit().unwrap();
 
@@ -766,7 +766,7 @@ mod test {
             let mut storage = InMemoryStorage::<Column>::default();
             let mut write = storage.write_transaction();
             let key = vec![0xA, 0xB, 0xC];
-            let expected = Arc::new(vec![1, 2, 3]);
+            let expected = Value::from([1, 2, 3]);
             write.put(&key, Column::Metadata, expected.clone()).unwrap();
             write.commit().unwrap();
 
@@ -787,7 +787,7 @@ mod test {
             let mut storage = InMemoryStorage::<Column>::default();
             let mut write = storage.write_transaction();
             let key = vec![0xA, 0xB, 0xC];
-            let expected = Arc::new(vec![1, 2, 3]);
+            let expected = Value::from([1, 2, 3]);
             write.put(&key, Column::Metadata, expected).unwrap();
             // test
             let ret = write.exists(&key, Column::Metadata).unwrap();
@@ -801,7 +801,7 @@ mod test {
             let mut storage = InMemoryStorage::<Column>::default();
             let mut write = storage.write_transaction();
             let key = vec![0xA, 0xB, 0xC];
-            let expected = Arc::new(vec![1, 2, 3]);
+            let expected = Value::from([1, 2, 3]);
             write.put(&key, Column::Metadata, expected).unwrap();
             write.commit().unwrap();
 
@@ -818,7 +818,7 @@ mod test {
             let mut storage = InMemoryStorage::<Column>::default();
             let mut write = storage.write_transaction();
             let key = vec![0xA, 0xB, 0xC];
-            let expected = Arc::new(vec![1, 2, 3]);
+            let expected = Value::from([1, 2, 3]);
             write.put(&key, Column::Metadata, expected).unwrap();
             write.commit().unwrap();
 
@@ -840,7 +840,7 @@ mod test {
             let mut storage = InMemoryStorage::<Column>::default();
             let mut write = storage.write_transaction();
             let key = vec![0xA, 0xB, 0xC];
-            let expected = Arc::new(vec![1, 2, 3]);
+            let expected = Value::from([1, 2, 3]);
             write.put(&key, Column::Metadata, expected.clone()).unwrap();
             // test
             write.commit().unwrap();
@@ -855,7 +855,7 @@ mod test {
             let mut storage = InMemoryStorage::<Column>::default();
             let mut write = storage.write_transaction();
             let key = vec![0xA, 0xB, 0xC];
-            let expected = Arc::new(vec![1, 2, 3]);
+            let expected = Value::from([1, 2, 3]);
             write.put(&key, Column::Metadata, expected).unwrap();
             write.commit().unwrap();
 
@@ -874,7 +874,7 @@ mod test {
 
             let mut storage = InMemoryStorage::<Column>::default();
             let mut write = storage.write_transaction();
-            let expected = Arc::new(vec![]);
+            let expected = Value::from([]);
             write.put(&key, Column::Metadata, expected.clone()).unwrap();
 
             assert_eq!(
@@ -912,7 +912,7 @@ mod test {
 
             let mut storage = InMemoryStorage::<Column>::default();
             let mut write = storage.write_transaction();
-            let expected = Arc::new(vec![]);
+            let expected = Value::from([]);
             write.put(&key, Column::Metadata, expected.clone()).unwrap();
 
             assert_eq!(
@@ -950,7 +950,7 @@ mod test {
 
             let mut storage = InMemoryStorage::<Column>::default();
             let mut write = storage.write_transaction();
-            let expected = Arc::new(vec![]);
+            let expected = Value::from([]);
             write.put(&key, Column::Metadata, expected.clone()).unwrap();
 
             assert_eq!(